Move from submodule to logger.py
